pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "echoscribe"
version = "2.0.0"
description = "Automated meeting transcription and summarization tool"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Hatim El Hassak" }
]
keywords = ["meeting", "transcription", "summarization", "slack", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "slack-sdk>=3.0.0",
    "openai>=1.0.0",
    "google-cloud-speech>=2.0.0",
    "pydub>=0.25.0",
    "sounddevice>=0.4.0",
    "scipy>=1.0.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
echoscribe = "echoscribe.cli:main"

[project.urls]
Homepage = "https://github.com/hatimhtm/EchoScribe"
Repository = "https://github.com/hatimhtm/EchoScribe"
Issues = "https://github.com/hatimhtm/EchoScribe/issues"

[tool.hatch.build.targets.wheel]
packages = ["echoscribe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=echoscribe --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
